Module tokio_core::channel

Deprecated since 0.1.1
: use futures::sync::mpsc instead

In-memory evented channels.

This module contains a Sender and Receiver pair types which can be used to send messages between different future tasks.

Structs

Receiver 	 [Deprecated]
The receiving half of a channel used for processing messages sent by a Sender.

Sender 	 [Deprecated]
The transmission half of a channel used for sending messages to a receiver.

Functions

channel 	 [Deprecated]
Creates a new in-memory channel used for sending data across Send + 'static boundaries, frequently threads.