Struct tokio_core::easy::EasyFramed

pub struct EasyFramed<T, P, S> { /* fields omitted */ }

An implementation of the FramedIo trait building on instances of the Parse and Serialize traits.

Many I/O streams are simply a framed protocol on both the inbound and outbound halves. In essence the underlying stream of bytes can be converted to a stream of frames. This way instead of reading or writing bytes a stream deals with reading and writing frames.

This struct is essentially a convenience implementation of the FramedIo which only requires knowledge of how to parse and serialize types. It is constructed with an arbitrary Io instance along with how to parse and serialize the frames that this EasyFramed will be yielding.

This implementation of FramedIo uses the EasyBuf type from the bytes crate for the backing storage, which should allow for zero-copy parsing where possible.

Methods

impl<T, P, S> EasyFramed<T, P, S> where T: Io, P: Parse, S: Serialize

fn new(upstream: T, parse: P, serialize: S) -> EasyFramed<T, P, S>

Creates a new instance of EasyFramed from the given component pieces.

This method will create a new instance of EasyFramed which implements FramedIo for reading and writing frames from an underlying I/O stream. The upstream argument here is the byte-based I/O stream that it will be operating on. Data will be read from this stream and parsed with parse into frames. Frames written to this instance will be serialized by serialize and then written to upstream.

The rd and wr buffers provided are used for reading and writing bytes and provide a small amount of control over how buffering happens.