Trait tokio_core::easy::Parse

pub trait Parse {
    type Out;
    fn parse(&mut self, buf: &mut EasyBuf) -> Poll<Self::Out, Error>;

    fn done(&mut self, buf: &mut EasyBuf) -> Result<Self::Out> { ... }
}

Implementation of parsing a frame from an internal buffer.

This trait is used when constructing an instance of EasyFramed. It defines how to parse the incoming bytes on a stream to the specified type of frame for that framed I/O stream.

The primary method of this trait, parse, attempts to parse a method from a buffer of bytes. It has the option of returning NotReady, indicating that more bytes need to be read before parsing can continue as well.

Associated Types

type Out

The type that this instance of Parse will attempt to be parsing.

This is typically a frame being parsed from an input stream, such as an HTTP request, a Redis command, etc.

Required Methods

fn parse(&mut self, buf: &mut EasyBuf) -> Poll<Self::Out, Error>

Attempts to parse a frame from the provided buffer of bytes.

This method is called by EasyFramed whenever bytes are ready to be parsed. The provided buffer of bytes is what's been read so far, and this instance of Parse can determine whether an entire frame is in the buffer and is ready to be returned.

If an entire frame is available, then this instance will remove those bytes from the buffer provided and return them as a parsed frame. Note that removing bytes from the provided buffer doesn't always necessarily copy the bytes, so this should be an efficient operation in most circumstances.

If the bytes look valid, but a frame isn't fully available yet, then Async::NotReady is returned. This indicates to the EasyFramed instance that it needs to read some more bytes before calling this method again.

Finally, if the bytes in the buffer are malformed then an error is returned indicating why. This informs EasyFramed that the stream is now corrupt and should be terminated.

Provided Methods

fn done(&mut self, buf: &mut EasyBuf) -> Result<Self::Out>

A default method available to be called when there are no more bytes available to be read from the underlying I/O.

This method defaults to calling parse and returns an error if NotReady is returned. Typically this doesn't need to be implemented unless the framing protocol differs near the end of the stream.