Module tokio_core::io

Deprecated
: moved to the tokio-io crate

I/O conveniences when working with primitives in tokio-core

Contains various combinators to work with I/O objects and type definitions as well.

A description of the high-level I/O combinators can be found online in addition to a description of the low level details.

Structs

Copy 	 [
Deprecated
]
A future which will copy all data from a reader into a writer.

EasyBuf 	 [
Deprecated
]
A reference counted buffer of bytes.

EasyBufMut 	 [
Deprecated
]
An RAII object returned from get_mut which provides mutable access to the underlying Vec<u8>.

Flush 	 [
Deprecated
]
A future used to fully flush an I/O object.

Framed 	 [
Deprecated
]
A unified Stream and Sink interface to an underlying Io object, using the Codec trait to encode and decode frames.

Read 	 [
Deprecated
]
A future which can be used to easily read available number of bytes to fill a buffer.

ReadExact 	 [
Deprecated
]
A future which can be used to easily read exactly enough bytes to fill a buffer.

ReadHalf 	 [
Deprecated
]
The readable half of an object returned from Io::split.

ReadToEnd 	 [
Deprecated
]
A future which can be used to easily read the entire contents of a stream into a vector.

ReadUntil 	 [
Deprecated
]
A future which can be used to easily read the contents of a stream into a vector until the delimiter is reached.

Window 	 [
Deprecated
]
A owned window around an underlying buffer.

WriteAll 	 [
Deprecated
]
A future used to write the entire contents of some data to a stream.

WriteHalf 	 [
Deprecated
]
The writable half of an object returned from Io::split.

Traits

Codec 	 [
Deprecated
]
Encoding and decoding of frames via buffers.

Io 	 [
Deprecated
]
A trait for read/write I/O objects

Functions

copy 	 [
Deprecated
]
Creates a future which represents copying all the bytes from one object to another.

flush 	 [
Deprecated
]
Creates a future which will entirely flush an I/O object and then yield the object itself.

read 	 [
Deprecated
]
Tries to read some bytes directly into the given buf in asynchronous manner, returning a future type.

read_exact 	 [
Deprecated
]
Creates a future which will read exactly enough bytes to fill buf, returning an error if EOF is hit sooner.

read_to_end 	 [
Deprecated
]
Creates a future which will read all the bytes associated with the I/O object A into the buffer provided.

read_until 	 [
Deprecated
]
Creates a future which will read all the bytes associated with the I/O object A into the buffer provided until the delimiter byte is reached. This method is the async equivalent to BufRead::read_until.

write_all 	 [
Deprecated
]
Creates a future that will write the entire contents of the buffer buf to the stream a provided.

Type Definitions

IoFuture 	 [
Deprecated
]
A convenience typedef around a Future whose error component is io::Error

IoStream 	 [
Deprecated
]
A convenience typedef around a Stream whose error component is io::Error