Struct tokio_core::io::ReadToEnd

pub struct ReadToEnd<A> {
    // some fields omitted
}

A future which can be used to easily read the entire contents of a stream into a vector.

Created by the read_to_end function.