Struct tokio_core::reactor::Handle

pub struct Handle { /* fields omitted */ }

A non-sendable handle to an event loop, useful for manufacturing instances of LoopData.

Methods

impl Handle

pub fn new_tokio_handle(&self) -> &Handle

Returns a reference to the new Tokio handle

pub fn remote(&self) -> &Remote

Returns a reference to the underlying remote handle to the event loop.

pub fn spawn<F>(&self, f: F) where
F: Future<Item = (), Error = ()> + 'static,

Spawns a new future on the event loop this handle is associated with.

Panics

This method will not catch panics from polling the future f. If the future panics then it's the responsibility of the caller to catch that panic and handle it as appropriate.

pub fn spawn_send<F>(&self, f: F) where
F: Future<Item = (), Error = ()> + Send + 'static,

Spawns a new future onto the threadpool

Panics

This function panics if the spawn fails. Failure occurs if the executor is currently at capacity and is unable to spawn a new future.

pub fn spawn_fn<F, R>(&self, f: F) where
F: FnOnce() -> R + 'static,
R: IntoFuture<Item = (), Error = ()> + 'static,

Spawns a closure on this event loop.

This function is a convenience wrapper around the spawn function above for running a closure wrapped in futures::lazy. It will spawn the function f provided onto the event loop, and continue to run the future returned by f on the event loop as well.

Panics

This method will not catch panics from polling the future f. If the future panics then it's the responsibility of the caller to catch that panic and handle it as appropriate.

pub fn id(&self) -> CoreId

Return the ID of the represented Core