Struct tokio_core::reactor::PollEvented

pub struct PollEvented<E> { /* fields omitted */ }

A concrete implementation of a stream of readiness notifications for I/O objects that originates from an event loop.

Created by the PollEvented::new method, each PollEvented is associated with a specific event loop and source of events that will be registered with an event loop.

Each readiness stream has a number of methods to test whether the underlying object is readable or writable. Once the methods return that an object is readable/writable, then it will continue to do so until the need_read or need_write methods are called.

That is, this object is typically wrapped in another form of I/O object. It's the responsibility of the wrapper to inform the readiness stream when a "would block" I/O event is seen. The readiness stream will then take care of any scheduling necessary to get notified when the event is ready again.

You can find more information about creating a custom I/O object online.

Methods

impl<E: Evented> PollEvented<E>

fn new(io: E, handle: &Handle) -> Result<PollEvented<E>>

Creates a new readiness stream associated with the provided loop_handle and for the given source.

This method returns a future which will resolve to the readiness stream when it's ready.

fn deregister(self, handle: &Handle) -> Result<()>

Deregisters this source of events from the reactor core specified.

This method can optionally be called to unregister the underlying I/O object with the event loop that the handle provided points to. Typically this method is not required as this automatically happens when E is dropped, but for some use cases the E object doesn't represent an owned reference, so dropping it won't automatically unreigster with the event loop.

This consumes self as it will no longer provide events after the method is called, and will likely return an error if this PollEvented was created on a separate event loop from the handle specified.

impl<E> PollEvented<E>

fn poll_read(&self) -> Async<()>

Tests to see if this source is ready to be read from or not.

If this stream is not ready for a read then NotReady will be returned and the current task will be scheduled to receive a notification when the stream is readable again. In other words, this method is only safe to call from within the context of a future's task, typically done in a Future::poll method.

fn poll_write(&self) -> Async<()>

Tests to see if this source is ready to be written to or not.

If this stream is not ready for a write then NotReady will be returned and the current task will be scheduled to receive a notification when the stream is writable again. In other words, this method is only safe to call from within the context of a future's task, typically done in a Future::poll method.

fn need_read(&self)

Indicates to this source of events that the corresponding I/O object is no longer readable, but it needs to be.

This function, like poll_read, is only safe to call from the context of a future's task (typically in a Future::poll implementation). It informs this readiness stream that the underlying object is no longer readable, typically because a "would block" error was seen.

The flag indicating that this stream is readable is unset and the current task is scheduled to receive a notification when the stream is then again readable.

Note that it is also only valid to call this method if poll_read previously indicated that the object is readable. That is, this function must always be paired with calls to poll_read previously.

fn need_write(&self)

Indicates to this source of events that the corresponding I/O object is no longer writable, but it needs to be.

This function, like poll_write, is only safe to call from the context of a future's task (typically in a Future::poll implementation). It informs this readiness stream that the underlying object is no longer writable, typically because a "would block" error was seen.

The flag indicating that this stream is writable is unset and the current task is scheduled to receive a notification when the stream is then again writable.

Note that it is also only valid to call this method if poll_write previously indicated that the object is writable. That is, this function must always be paired with calls to poll_write previously.

fn remote(&self) -> &Remote

Returns a reference to the event loop handle that this readiness stream is associated with.

fn get_ref(&self) -> &E

Returns a shared reference to the underlying I/O object this readiness stream is wrapping.

fn get_mut(&mut self) -> &mut E

Returns a mutable reference to the underlying I/O object this readiness stream is wrapping.