Struct tokio_core::LoopHandle

pub struct LoopHandle {
    // some fields omitted
}

Handle to an event loop, used to construct I/O objects, send messages, and otherwise interact indirectly with the event loop itself.

Handles can be cloned, and when cloned they will still refer to the same underlying event loop.

Methods

impl LoopHandle

fn channel<T>(self) -> (Sender<T>, IoFuture<Receiver<T>>) where T: Send + 'static

Creates a new in-memory channel used for sending data across Send + 'static boundaries, frequently threads.

This type can be used to conveniently send messages between futures. Unlike the futures crate channel method and types, the returned tx/rx pair is a multi-producer single-consumer (mpsc) channel with no backpressure. Currently it's left up to the application to implement a mechanism, if necessary, to avoid messages piling up.

The returned Sender can be used to send messages that are processed by the returned Receiver. The Sender can be cloned to send messages from multiple sources simultaneously.

impl LoopHandle

fn add_source<E>(&self, source: E) -> AddSource<E> where E: Evented + Send + 'static

Add a new source to an event loop, returning a future which will resolve to the token that can be used to identify this source.

When a new I/O object is created it needs to be communicated to the event loop to ensure that it's registered and ready to receive notifications. The event loop with then respond back with the I/O object and a token which can be used to send more messages to the event loop.

The token returned is then passed in turn to each of the methods below to interact with notifications on the I/O object itself.

Panics

The returned future will panic if the event loop this handle is associated with has gone away, or if there is an error communicating with the event loop.

fn schedule_read(&self, tok: &IoToken)

Schedule the current future task to receive a notification when the corresponding I/O object is readable.

Once an I/O object has been registered with the event loop through the add_source method, this method can be used with the assigned token to notify the current future task when the next read notification comes in.

The current task will only receive a notification once and to receive further notifications it will need to call schedule_read again.

Note: This method should generally not be used directly, but rather the ReadinessStream type should be used instead.

Panics

This function will panic if the event loop this handle is associated with has gone away, or if there is an error communicating with the event loop.

This function will also panic if there is not a currently running future task.

fn schedule_write(&self, tok: &IoToken)

Schedule the current future task to receive a notification when the corresponding I/O object is writable.

Once an I/O object has been registered with the event loop through the add_source method, this method can be used with the assigned token to notify the current future task when the next write notification comes in.

The current task will only receive a notification once and to receive further notifications it will need to call schedule_write again.

Note: This method should generally not be used directly, but rather the ReadinessStream type should be used instead.

Panics

This function will panic if the event loop this handle is associated with has gone away, or if there is an error communicating with the event loop.

This function will also panic if there is not a currently running future task.

fn drop_source(&self, tok: &IoToken)

Unregister all information associated with a token on an event loop, deallocating all internal resources assigned to the given token.

This method should be called whenever a source of events is being destroyed. This will ensure that the event loop can reuse tok for another I/O object if necessary and also remove it from any poll notifications and callbacks.

Note that wake callbacks may still be invoked after this method is called as it may take some time for the message to drop a source to reach the event loop. Despite this fact, this method will attempt to ensure that the callbacks are not invoked, so pending scheduled callbacks cannot be relied upon to get called.

Note: This method should generally not be used directly, but rather the ReadinessStream type should be used instead.

Panics

This function will panic if the event loop this handle is associated with has gone away, or if there is an error communicating with the event loop.

impl LoopHandle

fn add_timeout(&self, at: Instant) -> AddTimeout

Adds a new timeout to get fired at the specified instant, notifying the specified task.

fn update_timeout(&self, timeout: &TimeoutToken)

Updates a previously added timeout to notify a new task instead.

Panics

This method will panic if the timeout specified was not created by this loop handle's add_timeout method.

fn cancel_timeout(&self, timeout: &TimeoutToken)

Cancel a previously added timeout.

Panics

This method will panic if the timeout specified was not created by this loop handle's add_timeout method.

impl LoopHandle

fn spawn<F, R>(&self, f: F) where F: FnOnce(&LoopPin) -> R + Send + 'static, R: IntoFuture<Item=(), Error=()>, R::Future: 'static

Spawns a new future into the event loop this handle is associated this.

This function takes a closure which is executed within the context of the I/O loop itself. The future returned by the closure will be scheduled on the event loop an run to completion.

Note that while the closure, F, requires the Send bound as it might cross threads, the future R does not.

impl LoopHandle

fn tcp_listen(self, addr: &SocketAddr) -> IoFuture<TcpListener>

Create a new TCP listener associated with this event loop.

The TCP listener will bind to the provided addr address, if available, and will be returned as a future. The returned future, if resolved successfully, can then be used to accept incoming connections.

fn tcp_connect(self, addr: &SocketAddr) -> IoFuture<TcpStream>

Create a new TCP stream connected to the specified address.

This function will create a new TCP socket and attempt to connect it to the addr provided. The returned future will be resolved once the stream has successfully connected. If an error happens during the connection or during the socket creation, that error will be returned to the future instead.

impl LoopHandle

fn timeout(self, dur: Duration) -> IoFuture<Timeout>

Creates a new timeout which will fire at dur time into the future.

This function will return a future that will resolve to the actual timeout object. The timeout object itself is then a future which will be set to fire at the specified point in the future.

fn timeout_at(self, at: Instant) -> IoFuture<Timeout>

Creates a new timeout which will fire at the time specified by at.

This function will return a future that will resolve to the actual timeout object. The timeout object itself is then a future which will be set to fire at the specified point in the future.

impl LoopHandle

fn udp_bind(self, addr: &SocketAddr) -> IoFuture<UdpSocket>

Create a new UDP socket bound to the specified address.

This function will create a new UDP socket and attempt to bind it to the addr provided. The returned future will be resolved once the socket has successfully bound. If an error happens during the binding or during the socket creation, that error will be returned to the future instead.