Struct tokio_core::net::UdpSocket

pub struct UdpSocket { /* fields omitted */ }

An I/O object representing a UDP socket.

Implementations

impl UdpSocket

pub fn bind(addr: &SocketAddr, handle: &Handle) -> Result<UdpSocket>

Create a new UDP socket bound to the specified address.

This function will create a new UDP socket and attempt to bind it to the addr provided. If the result is Ok, the socket has successfully bound.

pub fn from_socket(socket: UdpSocket, handle: &Handle) -> Result<UdpSocket>

Creates a new UdpSocket from the previously bound socket provided.

The socket given will be registered with the event loop that handle is associated with. This function requires that socket has previously been bound to an address to work correctly.

This can be used in conjunction with net2’s UdpBuilder interface to configure a socket before it’s handed off, such as setting options like reuse_address or binding to multiple addresses.

pub fn framed<C: UdpCodec>(self, codec: C) -> UdpFramed<C>

Provides a Stream and Sink interface for reading and writing to this UdpSocket object, using the provided UdpCodec to read and write the raw data.

Raw UDP sockets work with datagrams, but higher-level code usually wants to batch these into meaningful chunks, called “frames”. This method layers framing on top of this socket by using the UdpCodec trait to handle encoding and decoding of messages frames. Note that the incoming and outgoing frame types may be distinct.

This function returns a single object that is both Stream and Sink; grouping this into a single object is often useful for layering things which require both read and write access to the underlying object.

If you want to work more directly with the streams and sink, consider calling split on the UdpFramed returned by this method, which will break them into separate objects, allowing them to interact more easily.

pub fn local_addr(&self) -> Result<SocketAddr>

Returns the local address that this stream is bound to.

pub fn connect(&self, addr: &SocketAddr) -> Result<()>

Connects the UDP socket setting the default destination for send() and limiting packets that are read via recv from the address specified in addr.

pub fn send(&self, buf: &[u8]) -> Result<usize>

Sends data on the socket to the address previously bound via connect(). On success, returns the number of bytes written.

pub fn recv(&self, buf: &mut [u8]) -> Result<usize>

Receives data from the socket previously bound with connect(). On success, returns the number of bytes read.

pub fn poll_read(&self) -> Async<()>

Test whether this socket is ready to be read or not.

If the socket is not readable then the current task is scheduled to get a notification when the socket does become readable. That is, this is only suitable for calling in a Future::poll method and will automatically handle ensuring a retry once the socket is readable again.

pub fn poll_write(&self) -> Async<()>

Test whether this socket is ready to be written to or not.

If the socket is not writable then the current task is scheduled to get a notification when the socket does become writable. That is, this is only suitable for calling in a Future::poll method and will automatically handle ensuring a retry once the socket is writable again.

pub fn send_to(&self, buf: &[u8], target: &SocketAddr) -> Result<usize>

Sends data on the socket to the given address. On success, returns the number of bytes written.

Address type can be any implementer of ToSocketAddrs trait. See its documentation for concrete examples.

pub fn send_dgram<T>(self, buf: T, addr: SocketAddr) -> SendDgram<T> where
T: AsRef<[u8]>,

Creates a future that will write the entire contents of the buffer buf provided as a datagram to this socket.

The returned future will return after data has been written to the outbound socket. The future will resolve to the stream as well as the buffer (for reuse if needed).

Any error which happens during writing will cause both the stream and the buffer to get destroyed. Note that failure to write the entire buffer is considered an error for the purposes of sending a datagram.

The buf parameter here only requires the AsRef<[u8]> trait, which should be broadly applicable to accepting data which can be converted to a slice. The Window struct is also available in this crate to provide a different window into a slice if necessary.

pub fn recv_from(&self, buf: &mut [u8]) -> Result<(usize, SocketAddr)>

Receives data from the socket. On success, returns the number of bytes read and the address from whence the data came.

pub fn recv_dgram<T>(self, buf: T) -> RecvDgram<T> where
T: AsMut<[u8]>,

Creates a future that receive a datagram to be written to the buffer provided.

The returned future will return after a datagram has been received on this socket. The future will resolve to the socket, the buffer, the amount of data read, and the address the data was received from.

An error during reading will cause the socket and buffer to get destroyed and the socket will be returned.

The buf parameter here only requires the AsMut<[u8]> trait, which should be broadly applicable to accepting data which can be converted to a slice. The Window struct is also available in this crate to provide a different window into a slice if necessary.

pub fn broadcast(&self) -> Result<bool>

Gets the value of the SO_BROADCAST option for this socket.

For more information about this option, see set_broadcast.

pub fn set_broadcast(&self, on: bool) -> Result<()>

Sets the value of the SO_BROADCAST option for this socket.

When enabled, this socket is allowed to send packets to a broadcast address.

pub fn multicast_loop_v4(&self) -> Result<bool>

Gets the value of the IP_MULTICAST_LOOP option for this socket.

For more information about this option, see set_multicast_loop_v4.

pub fn set_multicast_loop_v4(&self, on: bool) -> Result<()>

Sets the value of the IP_MULTICAST_LOOP option for this socket.

If enabled, multicast packets will be looped back to the local socket. Note that this may not have any affect on IPv6 sockets.

pub fn multicast_ttl_v4(&self) -> Result<u32>

Gets the value of the IP_MULTICAST_TTL option for this socket.

For more information about this option, see set_multicast_ttl_v4.

pub fn set_multicast_ttl_v4(&self, ttl: u32) -> Result<()>

Sets the value of the IP_MULTICAST_TTL option for this socket.

Indicates the time-to-live value of outgoing multicast packets for this socket. The default value is 1 which means that multicast packets don’t leave the local network unless explicitly requested.

Note that this may not have any affect on IPv6 sockets.

pub fn multicast_loop_v6(&self) -> Result<bool>

Gets the value of the IPV6_MULTICAST_LOOP option for this socket.

For more information about this option, see set_multicast_loop_v6.

pub fn set_multicast_loop_v6(&self, on: bool) -> Result<()>

Sets the value of the IPV6_MULTICAST_LOOP option for this socket.

Controls whether this socket sees the multicast packets it sends itself. Note that this may not have any affect on IPv4 sockets.

pub fn ttl(&self) -> Result<u32>

Gets the value of the IP_TTL option for this socket.

For more information about this option, see set_ttl.

pub fn set_ttl(&self, ttl: u32) -> Result<()>

Sets the value for the IP_TTL option on this socket.

This value sets the time-to-live field that is used in every packet sent from this socket.

pub fn join_multicast_v4(
&self,
multiaddr: &Ipv4Addr,
interface: &Ipv4Addr
) -> Result<()>

Executes an operation of the IP_ADD_MEMBERSHIP type.

This function specifies a new multicast group for this socket to join. The address must be a valid multicast address, and interface is the address of the local interface with which the system should join the multicast group. If it’s equal to INADDR_ANY then an appropriate interface is chosen by the system.

pub fn join_multicast_v6(
&self,
multiaddr: &Ipv6Addr,
interface: u32
) -> Result<()>

Executes an operation of the IPV6_ADD_MEMBERSHIP type.

This function specifies a new multicast group for this socket to join. The address must be a valid multicast address, and interface is the index of the interface to join/leave (or 0 to indicate any interface).

pub fn leave_multicast_v4(
&self,
multiaddr: &Ipv4Addr,
interface: &Ipv4Addr
) -> Result<()>

Executes an operation of the IP_DROP_MEMBERSHIP type.

For more information about this option, see join_multicast_v4.

pub fn leave_multicast_v6(
&self,
multiaddr: &Ipv6Addr,
interface: u32
) -> Result<()>

Executes an operation of the IPV6_DROP_MEMBERSHIP type.

For more information about this option, see join_multicast_v6.

pub fn set_only_v6(&self, only_v6: bool) -> Result<()>

Sets the value for the IPV6_V6ONLY option on this socket.

If this is set to true then the socket is restricted to sending and receiving IPv6 packets only. In this case two IPv4 and IPv6 applications can bind the same port at the same time.

If this is set to false then the socket can be used to send and receive packets from an IPv4-mapped IPv6 address.

pub fn only_v6(&self) -> Result<bool>

Gets the value of the IPV6_V6ONLY option for this socket.

For more information about this option, see set_only_v6.