Module tokio_core::reactor

The core reactor driving all I/O

This module contains the Core type which is the reactor for all I/O happening in tokio-core. This reactor (or event loop) is used to run futures, schedule tasks, issue I/O requests, etc.

Structs

Core
An event loop.

CoreId
An unique ID for a Core

Handle
A non-sendable handle to an event loop, useful for manufacturing instances of LoopData.

Interval
A stream representing notifications at fixed interval

PollEvented
A concrete implementation of a stream of readiness notifications for I/O objects that originates from an event loop.

Remote
Handle to an event loop, used to construct I/O objects, send messages, and otherwise interact indirectly with the event loop itself.

Timeout
A future representing the notification that a timeout has occurred.